Search SDK: Excluding some model types from code generation
SearchParameters and SuggestParameters will be maintained by hand from now on.
They are left in the spec so they can eventually be adapted to generate code
for the non-generic (i.e. -- dynamically-typed) parts of the Search and
Suggest APIs. The other models will eventually be generated again, once
AutoRest supports renaming of client properties. See:

https://github.com/Azure/autorest/issues/316