Removing begin* functionality and 307 logic
After working with PMs, we have decided to remove the 307 related logic
from our SDKs, as our SDKs are not required to be completely webhdfs
compliant. As a result, we can remove a lot of the complex post code gen
logic around redirect handling and multi level requests. This
drastically simplifies our SDK as well as our generation process.