Minor updates to ADL specs (#253)
* Remove requirement on subscriptionId

These clients don't need a subscriptionId.

* Need to make contents required since this is a PUT.

Will add extensions that make it optional in C# and zero it out to an
empty stream.

* update documentation to reflect param required

* undo this temp change

Issue: https://github.com/Azure/autorest/issues/884 is tracking support
for this being an optional param properly and not resulting in a null
reference. So I will keep this spec as is.

* Change this name to be more descriptive.

* does not need to be modeled as a string.

Makes this more obvious to end users and easier to understand.

* fix all enums to be enums.

* Add new API version for job

This API version changes the format for dates and timespan objects to be
ISO compliant.

* missed version change here as well.

* temp commit to remove the required flag

We need a way to have template parameters not be required when they are
in the host. I will work with Amar for this next week.

* Revert "temp commit to remove the required flag"

This reverts commit be8b7bf40e56c8316f423a871ccf030efe47a8a1.

* Add missing enum value

* Missed one.

* Add in new functionality

Switch over to MSLISTSTATUS which gives more flexibility for listing
files/folders.

* Work around for issue: 903

explicitly calling out our chunked transfer APIs until issue:
https://github.com/Azure/autorest/issues/903 is resolved.

* addressed comments from Amar.