Update ADL catalog (#346)
* Remove requirement on subscriptionId

These clients don't need a subscriptionId.

* Need to make contents required since this is a PUT.

Will add extensions that make it optional in C# and zero it out to an
empty stream.

* update documentation to reflect param required

* undo this temp change

Issue: https://github.com/Azure/autorest/issues/884 is tracking support
for this being an optional param properly and not resulting in a null
reference. So I will keep this spec as is.

* Change this name to be more descriptive.

* does not need to be modeled as a string.

Makes this more obvious to end users and easier to understand.

* fix all enums to be enums.

* Add new API version for job

This API version changes the format for dates and timespan objects to be
ISO compliant.

* missed version change here as well.

* temp commit to remove the required flag

We need a way to have template parameters not be required when they are
in the host. I will work with Amar for this next week.

* Revert "temp commit to remove the required flag"

This reverts commit be8b7bf40e56c8316f423a871ccf030efe47a8a1.

* Add missing enum value

* Missed one.

* Add in new functionality

Switch over to MSLISTSTATUS which gives more flexibility for listing
files/folders.

* Work around for issue: 903

explicitly calling out our chunked transfer APIs until issue:
https://github.com/Azure/autorest/issues/903 is resolved.

* addressed comments from Amar.

* fix descriptions and add partition APIs

* Add missing method to delete all secrets in db

In hyak, we could do this by simply excluding the parameter. That is not
swagger compliant, so we need a separate entry.

* temporarily remove partition list for current ship.

this will be in the next snap.

* Add diagnostics to USqlJobProperties.

This was missing during the port from Hyak.

* Missing enum value.

* two more missing enums.

* Put partitions back in for next release.

* Test out composites.

* add the log folder and logFilePattern to the sdk

This is for all jobs going forward.
Fix the readOnly properties to be bool instead of strings (makes tests
happy).

* Remove composite specs since they don't work

* Error json fixes

InnerError fixed to innerError
Adding the WebHDFS definition for ErrorResponses to the filesystem.

* bad syntax is bad.

* missed a boolean switch for readOnly,.

* definition -> definitions

* Change the name to be more descriptive

* Following feedback from the slack forum

This should help reduce the levels of nesting and also fixes some
feedback in the job swagger.

* Add discriminators for webhdfs

* Further flatten and rename

Removing WebHdfsError and renaming "RemoteException"

* Rename to Adls

* nested exception discriminator

* update the required piece.

* try not indexing

* move it down to see if this fixes it.

* Try nesting again.

* Trying a work around.

Add a placeholder property at the root level which should get properly
picked up.

* Update to the work around

Slight change to the stub property.

* Work around didn't work, going back to what did.

* Removed the unneccessary nesting.

* Add the accesscontrolexception

* remove extra next link stuff

I thought I had removed this previously. removing it again, since it
should no longer be used.

* Remove currently unsupported API

* Adding missing x-ms-odata tags in catalog

* switch to GET for this call per the FE team

This API supports both a GET and a POST, with the POST planning on being
deprecated before we GA, so we are switching the SDK now in preparation.

* Add partition properties

parentName
createDate

* missed the #

* Removing get/set from descriptions

latest generation tool automatically adds it.

* Fixed to not remove "Gets " from method descript

* Add the get back to globals.

* Catalog item updates

Add distributionInfo to tables
Add a new type called USqlTableType

* Fix URL based on feedback

should be tabletypes instead of tables