ADL Update for catalog and filesystem (#391)
* Need to make contents required since this is a PUT.

Will add extensions that make it optional in C# and zero it out to an
empty stream.

* update documentation to reflect param required

* undo this temp change

Issue: https://github.com/Azure/autorest/issues/884 is tracking support
for this being an optional param properly and not resulting in a null
reference. So I will keep this spec as is.

* temp commit to remove the required flag

We need a way to have template parameters not be required when they are
in the host. I will work with Amar for this next week.

* Revert "temp commit to remove the required flag"

This reverts commit be8b7bf40e56c8316f423a871ccf030efe47a8a1.

* Work around for issue: 903

explicitly calling out our chunked transfer APIs until issue:
https://github.com/Azure/autorest/issues/903 is resolved.

* Add missing method to delete all secrets in db

In hyak, we could do this by simply excluding the parameter. That is not
swagger compliant, so we need a separate entry.

* Test out composites.

* Remove composite specs since they don't work

* update the required piece.

* Try nesting again.

* Removing get/set from descriptions

latest generation tool automatically adds it.

* Fixed to not remove "Gets " from method descript

* Add the get back to globals.

* make typefieldinfo an array.

* Adding offset support for append operations

This is an optional value used for unique implementations that know
exactly where in the stream they want to begin appending data.

* Add file already exists exception for webhdfs

* Add BadOffsetException to swagger.

* Fix issues with bad rebase

had to copy back from azure and make the minor edits to these two .json
files.