Search SDK: Removing unnecessary analysis components
Some analyzers, tokenizers, token filters, and character filters do not take
any parameters. For this reason, it is confusing to leave their corresponding
types in the API surface area when it is sufficient to refer to them by name.

Also fixed some other documentation and changed lightItalian to
light_italian to work around a bug in the API (fix coming soon).