Search SDK: Merging all v2.0-preview feature work to master branch (#469)
* Search SDK: Disabling synchronous methods for SearchIndexClient

Currently we are using the spec for SearchIndexClient to generate a proxy that
is then wrapped with hand-written code. The hand-written C# code always calls
async methods, so there is no need for the proxy to have synchronous methods.

* Search SDK: Modeling AnalyzerName and DataType as strong types

AnalyzerName and DataType are like enums, but they need to be open-ended, so
we have a custom type for them. It is better than a class of static string
properties since it facilitates type-safety and Intellisense (for .NET code).

* Search SDK: Adding new indexer features to 2.x-preview spec

* Search SDK: Turning off codegen for IndexingParameters

We need to mark a property of this class with the [Obsolete] attribute. We
need to do this by hand until this AutoRest issue is fixed:

https://github.com/Azure/autorest/issues/1285

* Search SDK: Adding field mappings to API spec

* Search SDK: Modifying description of DataContainer.query

It now takes into account the existence of new datasource types.

* ETags for Azure Search Indexes, Indexers, and DataSources :mag_right:

* Search SDK: Adding support for custom analyzers

* Search SDK: Adding descriptions to some external types