Enable Encryption support for ADLS
encryptionState is not, in fact, read only

Change delete to a post to match the new schema

This is required to allow the user to pass in the password if necessary.

follow the existing pattern for body params

Missed description and a slash.

fix up the credential swagger and add descriptions.