Updating Creating Swagger doc code blocks
Since JSON is very strict about comments and non-JSON characters, Github does some ugly highlighting when using `...` to represent omitted lines in our documentation. By switching to the `json5` hint for large code blocks, it's a little easier to read.

This is only a documentation formatting change. It has no impact on the specs in this repo.