Update swagger-authoring-descriptions.md
Looks great! Hope you'll forgive the copy-editing suggestions.

#7 in the list could be "Include links to other content when necessary," and point to how/where to find those. I don't know whether these descriptions support links, so I'll leave that to you to include if it's worthwhile. 

You might also note that it's helpful to indicate in the description for a request header/query parameter whether it is required or optional.

I'm not sure what "the Linter" is, although I heard you say it the other day, so that might be good to define here.