Removed Erroneous SubscriptionID as Required Path Parameter
The generated code from this Azure spec is currently failing with the error shown in this ticket: 

https://github.com/Azure/autorest/issues/1550

We determined that the problem was that the subscriptionID parameter should not be a required path parameter.  The subscriptionID is already included in the resourceUri parameter, so it was clearly redundant and also causing calls for metrics to fail.  Only after we removed the "subscriptionId" field as a required parameter in the generated sources, were we able to retrieve metric definitions with the generated code. 

The same problem exists on the "metricDefinitions" endpoint on api version 2016-03-01.  I have submitted the following PR for that:  
https://github.com/Azure/azure-rest-api-specs/pull/660