Removing x-ms-azure-resource tag for ADL subs (#811)
This is causing problems in Ruby and Java where they assume a very
specific set of properties for sub resources (only an ID) and all
additional properties are ignored. This causes breaking problems for us.
There is a separate thread tracking this for a complete resolution. In
the meantime, I do not have a need to tag my sub resources as
x-ms-azure-resource.