Update allowed enum values for severity (#844)
* Update allowed enum values for severity

Updating the allowed enum values to include SevereWarning, Deprecated
and UserWarning, which the service has been able to return but were not
incorporated properly in the client.

* Update filesystem.json

Clarifying helptext for ConcurrentAppend, Append, and Create.

* Bug fix for updating secrets in ADL

Missed this when fixing credentials. No objects are returned when
creating or updating secrets or credentials.