Addressing WebHDFS team feedback about APIs
These are changes that are necessary in order to properly match the
Front end contract, since our swagger specs are designed only for client
construction, it is an ongoing process to bring them in line with the
current state of the service.

missed explicit required false.

Filesystem updates

Add leaseId
Add FileSessionId
Switch everything to not use MS* apis except for MSConcat.

Address CR Feedback

add optional "permission" to the CREATE calls
Explicitly call out that APPEND is for serial appends.

Address CR comments

Include tooId for GET operations (files and acls)
Add aclBit as a property that is returned in fileStatus
fixed a copy/paste comment.