Add backupConfiguration to Analysis services request/response properties (#946)
* Add backupConfiguration to Analysis services request/response properties

This change is to support backup when create, update or get server
details. backup configuration has 3 properties, storage account and blob
container are required for all operations. Access key is required for
create and update operations.

* Update the description for storageAccount field of BackupConfiguration

This change is to update the description for storageAccount field of
BackupConfiguration. Before GA, storageAccount will be updated to use
full storage acocunt resource id.