ADLA Catalog updates (#1119)
* ADLA Catalog updates

Inclusion of USqlPackages
(BREAKING) Switch FileType enum to be modeled as a string. This is
important since this field changes frequently and we don't want to break
the SDK for this read-only property
Add support for listing various catalog elements from higher level
nodes:
tables, views, statistics and tvfs from within a database (no schema
required)
statistics from within a schema (no table required).

* Update to add examples for new functionality


* Adding next links to examples

Note that these are "fake" next link values that may be different in a
real request, but they will be followable.