Swagger specs for Stream Analytics
- Added swagger specs for Azure Stream Analytics 2016-03-01 API version.
This includes management APIs for the top-level tracked resource
"streamingjobs", the management APIs for the nested/child resources
"inputs", "outputs", "transformations", and "functions" which are under
a streaming job as well as subscription level APIs to get the current
quotas for the subscription with respect to Azure Stream Analytics.
- Added x-ms-example files for all of the APIs defined in the swagger
specs. All of these examples should be referenced in the appropriate
swagger spec file.