Creating a new API version
* We decided to create a new API version (2017-09-01-preview) because we
found out that this Scope on ADL project introduces a REST-API breaking
change.  That is, we are adding the Scope type via the type field, which
is an enum.  Therefore, if an existing user, who has not upgraded to the
new SDK, gets a list of jobs that includes a Scope job, then the client
would crash because it would not know how to de-serialize the Scope enum
string.

* Thus, I added a new directory named 2017-09-01-preview and moved over
my job Swagger file with the Scope changes in there.  As for the job
Swagger file in 2016-11-01, I made sure to revert back to the old
version with some minor changes for descriptions.
* I made sure to update the version field in the info object of the job
Swagger file to be "2017-09-01-preview"