Refactor to host specs in preview/stable folders. (#2167)
This is in an effort to differentiate more clearly the Open API Specs
that are ready for public consumption, and those that should be treated
as a preview or in development. Simply treating the presence of the
suffux `-preview` on an API Version is insufficient, because the status
of the Swagger itself may be preview, while the API Version is release
quality.