Removing duplicate ResourceHealth files.
I've confirmed that the deleted files in the `examples` folder are duplicates of their counterparts in the `specification` directory.

Also, it seems that `resourcehealth.json` replaced a path variable for "resource type" with a static definition. This makes sense, as the Swagger documents themselves refer to exactly one resource type.