[Azure Search] Updating AutoRest config files
Fixed client name in the management plane config.

Split the data plane config into three files:
  - readme.md: Contains the config settings common to both data plane specs.
  - readme.index.md: Contains config settings specific to SearchIndexClient.
  - readme.service.md: Contains config settings specific to
SearchServiceClient.

This was necessary in order to get the Swagger validator to run correctly,
since otherwise it was combining the specs into one and crashing because they
both contain x-ms-parameterized-host. The intent is to have separate specs,
since the Azure Search data plane SDK is comprised of two separate clients
(one statically-typed for the base resources, and one dynamically-typed for
customer-specific index schemas).