Generate the latest package
The latest package including support for the new API will be generated.
This required several changes:
- The defintion of Resource used by the other specs was incompatible.
This was because the other resources were tracked and the new ones is
proxy. Defined a new ProxyResource type.
- The ErrorResponse definition was incompatible with the rest of the
specs in the folder. This is actually a bug in the exesiting specs. They
do not respect the contract of  { "error": {"code": "", "message": ""}},
i.e., they lack the top-level error element. To work around this issue,
defined a new response type, ResponseWithError, which is compliant
with the latest ARM requirements.