Add version 2018-11-27-preview to Microsoft.Insights (#4987)
* Adds base for updating Microsoft.Insights from version preview/2018-06-01-preview to version 2018-11-27-preview

* Updates readme

* Updates API version in new specs and examples

* Add initial swagger

Removed some of the autogenerated files and added the swagger file and
examples for the new api.

* Remove extra / from examples

* Fix swagger file

* Improve docs

* Update resource model

Made it the same as the rest of the specs

* Fix the resource data model: no location and tags

* Fix readme

* Generate the latest package

The latest package including support for the new API will be generated.
This required several changes:
- The defintion of Resource used by the other specs was incompatible.
This was because the other resources were tracked and the new ones is
proxy. Defined a new ProxyResource type.
- The ErrorResponse definition was incompatible with the rest of the
specs in the folder. This is actually a bug in the exesiting specs. They
do not respect the contract of  { "error": {"code": "", "message": ""}},
i.e., they lack the top-level error element. To work around this issue,
defined a new response type, ResponseWithError, which is compliant
with the latest ARM requirements.

* Update Python version