[Azure Search] Add transform directives for C# generation
This applies only to the Microsoft.Azure.Search.Data package, the source for
which is in azure-sdk-for-net.

We'll be using transform directives more often in the future, so it seemed
like a good idea to port our current hacky PowerShell script to this new
mechanism. The PowerShell script lives in azure-sdk-for-net, and will be
removed in a future PR.