[Azure Search] Add Get Documents (a.k.a. Lookup) API to data plane spec
Note that this API already exists in the .NET SDK, but is currently
implemented as custom code. We need to add transform directives for the
generated code because otherwise there's no way to pass custom
JsonSerializerSettings into the generated code (which is a requirement for
our data plane).

Also renamed some example files to be more consistent with existing naming
conventions.