[Azure Search] Updating some Swagger descriptions to be more doc-friendly
Sometimes we over-describe things, especially "comma-separated lists" in
query string parameters. These are modeled as arrays in Swagger, which are
already understood to be comma-separated lists. They are modeled as
collections in generated code, so they are decidedly not comma-separated
lists in that context. So removing this phrase is a win-win.