[Azure Search] Make Swagger spec accurately reflect the semantics of Field
Before this change, the descriptions in the Swagger for the Field type did
not accurately describe the semantics of the Azure Search REST API,
particularly around what the default values of the searchable, filterable,
sortable, and facetable properties were. Instead, they reflected the defaults
of our custom implementation of Field in the .NET SDK.

This change corrects the descriptions to accurately reflect the REST API. It
also adds more detail in anticipation of generating REST API documentation
from the Swagger (which we don't currently do).

A few other changes to Field are related to code generation:
 - Removed the x-ms-external flag so that AutoRest can generate this type.
 - Removed x-nullable: false and x-ms-client-name since they again reflect
   only the custom .NET code and not the desired REST API semantics.
 - Added regex transform rules to readme.md so that we can keep some of our
   custom code for Field in .NET purely for backward compatibility purposes.
   Our intention going forward for other languages is to rely solely on code
   generation for the Field type.