Adding the Microsoft.Security ServerVulnerabilityAssessment API.
ServerVulnerabilityAssessment is a new feature in Azure Security Center, where
Azure clients will be able to onboard to ASC Server Vulnerability Assessment
solution, and GET VA findings on their servers.

In order to onboard, a PUT request will be triggered on a VM server:
- PUT {resourceId}/providers/Microsoft.Security/serverVulnerabilityAssessments/default

At any point the user will be able to GET the onboarding status for his VM:
- GET {resourceId}/providers/Microsoft.Security/vulnerabilityAssessments/default

So if the VM is onboarded, a 200 response will be returned, and 404 otherwise.

For resources that are not VMs, a 405 "Not allowed" response will return.

Actual VA Findings will be fetched using ASC standard assessment API.