Adding new path-based rule for the Compute API
This is a new yml rule that specifies a configuration helpful to the cplat (Compute Platform) sdk team. This configuration rule mainly specifies a path, here the Compute Management library's path, and an email that will be sent by the bot whenever a PR is opened and the PR changes are in the Compute Management library's path.