Add "mail" parameter to UserUpdateParameters in graphrbac
This patch adds a single optional field "mail" to UserUpdateParameters
schema. This makes it possible to update user's e-mail address using
the (Go) Azure SDK. Currently this is not possible.

I have curled the graphrbac API with the extra body parameter and it
works as expected.

Addition of the field will make it possible to support the mail property
on the azuread_user resource in terraform-provider-azuread.