Fix parent class of ClusterResource and DataCenterResource in .NET SDK (#13244)
* Fix parent class of ClusterResource and DataCenterResource in .NET SDK

Removed x-ms-azure-resource from ClusterResource and DataCenterResource. x-ms-azure-resource is already set on the ProxyResource and TrackedResource definitions that ClusterResource and DataCenterResource, and when it's applied it on ClusterResource or DataCenterResource directly, it forces the parent class to be IResource in the .NET SDK and loses the location, type, id, etc. parameters.

* Switched to CosmosDB's ARM resource parents.

The types.json ones throw errors. In particular, I can't use TrackedResource because it sets location to be required, and the same field is not found by the linter.

* Remove required properties since they are forbidden for PATCH.

* Remove x-ms-client-flatten to avoid breaking changes.