Start moving managed Cassandra out from cosmos-db's Swagger definitions.
Managed Cassandra has only been in preview so far, so we can still do this without a breaking change in the stable API.
CosmosDB's resource provider is now large enough where multiple features are getting put into preview and then rolled into
stable on different schedules. One feature gets brought to stable, but then the next preview version has to merge stable plus
the changes from the previous preview version that are not in stable yet. This is getting harder and harder.

Since managed Cassandra is basically independent of the rest of the CosmosDB swagger, we are going to move out
to a new location so we can version ourselves separately.