[Purview Accounts] SDK Team Feedback on Operation IDs
This change updates the operation ids for operations in the Purview
Accounts service. These names are used when generating clients and so
for some we've added suffxes (based on their groupings) to make
generated method names a little clearler when you see just the method
name in code e.g.:

```
var client = new PurviewAccountsClient();
...
client.GetAccountProperties() // Nicer than just client.Get();
```

In addition, we noticed that the `Collections_GetCollectionPath`
operation took a `$skipToken` parameter, even though the operation is
not pageable. After following up with the service team we learned this
was a mistake, so I have removed the parameter from the specification.