Use "5Qi" as the schema to define Default QoS Flow 5G QoS Indicator value. (#9817)
On SIM policy we use the 'Default QoS Flow 5G QoS Indicator value', we were incorrectly using the '5QiPriorityLevel' schema (which is an interger between 1 and 127 as per https://github.com/Azure/azure-rest-api-specs-pr/blob/RPSaaSMaster/specification/mobilenetwork/resource-manager/Microsoft.MobileNetwork/stable/2022-11-01/ts29571.json#L117. We should have used 5Qi - an integer between 0 and 255. We made a mistake when we first p ut this in and have no spotted it.

Even though changing the allowed values for a property is a breaking change, we think this is acceptable because:
1) This is our first stable API, there is no pre-existing stable API we need to be back-compatible with
2) 2022-04-01-preview to 2022-11-01 already contains breaking changes that we have discussed in https://eng.ms/docs/strategic-missions-and-technologies/strategic-missions-and-technologies-organization/azure-for-operators/packet-core/private-mobile-network/azure-private-5g-core/cross-team/releases/2022-11-01-arm-api/back-compatibility
3) We have one customer who may need to use values 128-255 -> if that customer does PUTs with the 2022-11-01 API then a GET with the 2022-04-01, it will fail, while this is sad, we think this is an acceptable risk.