The run.sh script requires git, but git is not available by default in the base image
I've created a shell in the datacontainer and noticed that the run.sh script fails, because the git command is not found.  Without it, the datacontainer cannot check out the local clone of the repo and we end up with a completely empty/usr/data/openapispecs. This is currently blocking RPaaS api-validation service in Air-Gapped clouds, where we depend entirely on specs loaded from the data container.

I tested this change locally: WIth git being previously installed to the image, run.sh succeeds and /usr/data/openapispecs is properly populated with the repo contents.