Add specification for AVD personal schedules
Add the specification and examples for Azure Virtual Desktop personal scaling plan schedules.
Operations added: Create, Delete, Update, List (all schedules assigned to specified scaling plan).

Modified the CapacityThresholdPct in two places for scaling plans, to better align with the other places and what is already done in the backend. The previous setting minimum does not make sense and is changed in the service to the new minimum.