{AzureNetwork} fixes Azure/azure-rest-api-specs#22448 Making priority required
fixes Azure/azure-rest-api-specs#22448

[This is the API](https://learn.microsoft.com/en-us/rest/api/virtualnetwork/network-security-groups/create-or-update?tabs=HTTP#code-try-0) I invoked.

For the NSG security rules without priorities, seems like the priority parameter is required. Azure portal works as expected.
I don't think that's allowed but according to the schema it is because the priority property is not marked as required. I tried to create a NSG security rule without the priorities and it failed with below. Intrestingly the Azure 

Request Body:
```
{
  "properties": {
    "securityRules": [
      {
        "name": "rule1",
        "properties": {
          "protocol": "*",
          "sourceAddressPrefix": "*",
          "destinationAddressPrefix": "*",
          "access": "Allow",
          "destinationPortRange": "80",
          "sourcePortRange": "*",
          "direction": "Inbound"
        }
      }
    ]
  },
  "location": "eastus"
}
```
Response received:

```
{
  "error": {
    "code": "SecurityRuleInvalidPriority",
    "message": "Security rule has invalid Priority. Value provided: 0 Allowed range 100-4096.",
    "details": []
  }
}
```

This PR will make the `priority` parameter as required.