Add Personal HostPool Scaling
This change adds personal host pool scaling to Desktop Virtualization.
It allows for the users to Get, list, add, update, and delete schedules for scaling plans that manage personal host pools.
This change also changes the minimum value for rampDownCapacityThresholdPtc in Pooled to 1 in order to align with the minimum enforced in the backend.