Fix permissions for workflow
Need to fix permissions for the workflow to be able to successfully use gh pr edit as it needs to also be able to read the project list. This is only an issue in the private repo but we need we can set this up to work on either public or private.