Handle cases where the there are no changed files
In PS using "| Where-Object" on a null acts like
there is one null object in the list so trying
access properties on it fails. Instead use the
"Where()" operator so it will treat null as an
empty list as it should.
 This is a combination of 4 commits.

Update ChangedFiles-Functions.ps1

Update ChangedFiles-Functions.ps1

Update ChangedFiles-Functions.ps1

Update ChangedFiles-Functions.ps1