Introduce SharingScope property on Subnet
SharingScope property on a subnet is used to indicate if the subnet can be shared for Injection. Customer can set the enum value to 'Tenant' to start creating compute (VM/VMSS/Flex) on different subscriptions (under the same AAD tenant) while consuming IP Address from the subnet hosted on a different subscription.