Update ResourceProvider.json
hot fix

API will return with the correct body but also with a nextLink with a URL instead of just null. I did a little more research and it seems like since my API does not return next links I think the fix should just be set the nextlinkname to null in the swagger specification.