fix: remove 204 status code when completing blue/green ugprades
Addresses the following linter error:
Long-running POST operations must have responses with 202 and default return codes. They must also have a 200 return code if only if the final response is intended to have a schema, if not the 200 return code must not be specified. They also must not have other response codes.