[Microsoft.Subscription] Removes the 2016-06-01 API
Removes the 2016-06-01 API from the Microsoft.Subscription service. This API never should have been included with this service specification as it redefined apis from the Microsoft.Resources services (like `/tenants` and `/subscriptions`). 

This commit also updates the default tag for the service to `2024-08-01-preview` to mirror the actual evolution of the API.